refactor(api): split contexts.js controller into per-responsibility modules (#599)
contexts.js had grown to 637 lines (audit finding C1). Split it by
responsibility into focused routers under routes/contexts/:

  read.js    — GET /api/contexts (+ /tree, /:id, /:id/members) + loadMembers
  crud.js    — POST /api/contexts + PATCH/POST-sync/DELETE /contexts/:id
  members.js — POST/DELETE/PATCH-move /contexts/:id/members[...]
  shared.js  — useSql / UUID_RE / TARGET_TYPES + the writeContexts gate

contexts.js is now a barrel that composes the three routers, so app.js
and every public path are unchanged. Pure code move.

Grandfathered complexity entries relocated (no increase): cyclomatic
[30,22,21] -> crud [30,22] + members [21]; cognitive [40] -> crud.

No new tests needed — the existing contexts tests cover the moved code at
96%+ diff coverage, clearing the 80% gate. Full API suite green (1758
tests); coverage above every floor; both complexity gates and the
OpenAPI-drift guard pass.
