+++ .ci/cognitive-baseline.json
@@ -72,7 +72,7 @@
     "app/api/src/routes/authRoles.js": [
       16
     ],
-    "app/api/src/routes/contexts.js": [
+    "app/api/src/routes/contexts/crud.js": [
       40
     ],
     "app/api/src/routes/crawlers.js": [
