+++ .ci/complexity-baseline.json
@@ -34,9 +34,11 @@
     "app/api/src/routes/admin/curatedData.js": [
       30
     ],
-    "app/api/src/routes/contexts.js": [
+    "app/api/src/routes/contexts/crud.js": [
       30,
-      22,
+      22
+    ],
+    "app/api/src/routes/contexts/members.js": [
       21
     ],
     "app/api/src/routes/details/accessPackage.js": [
